% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrateSSB.R
\name{CalibrateSSBpanel2}
\alias{CalibrateSSBpanel2}
\title{Calibration weighting and variance estimation for panel data}
\usage{
CalibrateSSBpanel2(
  data,
  ...,
  numerator,
  denominator = NULL,
  linComb = NULL,
  linComb0 = NULL,
  estType = "robustModel",
  leveragePower = 1/2,
  group = NULL,
  returnCov = FALSE,
  usewGross = TRUE
)
}
\arguments{
\item{data}{The gross sample data frame}

\item{...}{Further arguments to CalibrateSSB including variables needed by WideFromCalibrate}

\item{numerator}{Argument to PanelEstimation}

\item{denominator}{Argument to PanelEstimation}

\item{linComb}{Argument to PanelEstimation}

\item{linComb0}{Argument to PanelEstimation}

\item{estType}{Argument to PanelEstimation}

\item{leveragePower}{Argument to PanelEstimation}

\item{group}{Argument to PanelEstimation}

\item{returnCov}{Argument to PanelEstimation}

\item{usewGross}{Argument to PanelEstimation}
}
\value{
Output from PanelEstimation combined into a data frame
}
\description{
The function is similar to \code{\link{CalibrateSSBpanel}} but rewritten since the advanced original code is problematic to re-export.  
Output is combined into a data frame. The function combines 
\code{\link{CalibrateSSB}}, \code{\link{WideFromCalibrate}} and \code{\link{PanelEstimation}}
into a single function.
}
\examples{
z    <- AkuData(3000)  # 3000 in each quarter
zPop <- AkuData(10000)[,1:7]
lc <- rbind(CalibrateSSB::LagDiff(8,4),CalibrateSSB::PeriodDiff(8,4))
rownames(lc) <- c("diffQ1","diffQ2","diffQ3","diffQ4","diffYearMean")
CalibrateSSBpanel2(z, calmodel="~ sex*age", partition=c("year","q"), popData=zPop, 
       y=c("unemployed","workforce"), id="id", wave=c("year","q"),
       numerator="unemployed", linComb=lc)
       
}
