% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrateSSB.R
\name{CalibrateSSBweights}
\alias{CalibrateSSBweights}
\title{Calibration weighting}
\usage{
CalibrateSSBweights(data, ..., weightsname = "calWeights")
}
\arguments{
\item{data}{The gross sample data frame}

\item{...}{Further arguments to CalibrateSSB}

\item{weightsname}{Name of variable with output weights}
}
\value{
Input data extended with output weights
}
\description{
A wrapper of the function  \code{\link{CalibrateSSB}}
}
\examples{

# Generates data  - two years
z    <- AkuData(3000)  # 3000 in each quarter
zPop <- AkuData(10000)[,1:7]

# Calibration using "survey"
a <- CalibrateSSBweights(z, calmodel = "~ sex*age",
                 partition = c("year","q"),  # calibrate within quarter
                 popData = zPop)

}
