% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallCountRounding.R
\name{PLSroundingPublish}
\alias{PLSroundingPublish}
\title{PLS inspired rounding - data frame output}
\usage{
PLSroundingPublish(data, ...)
}
\arguments{
\item{data}{Input data as a data frame (inner cells)}

\item{...}{Further arguments to PLSrounding}
}
\value{
Data frame of publishable data with the main dimensional variables and with cell frequencies (original, rounded, difference).
}
\description{
A wrapper of the function  \code{\link{PLSrounding}} in package SmallCountRounding
with \code{publish} as the only output.
}
\examples{
z <- SmallCountData("e6")
PLSroundingPublish(z, "freq", roundBase = 5,  formula = ~geo + eu + year)

}
