% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easySdcTable.R
\name{ProtectTableData}
\alias{ProtectTableData}
\title{Table suppression according to a frequency rule}
\usage{
ProtectTableData(data, ...)
}
\arguments{
\item{data}{The gross sample data frame}

\item{...}{Further arguments to ProtectTable}
}
\value{
Data set output from ProtectTable
}
\description{
A wrapper of the function  \code{\link{ProtectTable}} in package easySdcTable
}
\examples{

z1 <- EasyData("z1") 
ProtectTableData(z1, c("region","hovedint"), "ant",  method="SIMPLEHEURISTIC")

z1w <- EasyData("z1w") 
ProtectTableData(z1w, 1, 2:5)    
}
