% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrateSSB.R, R/SSBtools.R,
%   R/SmallCountRounding.R, R/easySdcTable.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{AkuData}
\alias{SSBtoolsData}
\alias{HierarchyCompute}
\alias{SmallCountData}
\alias{EasyData}
\alias{ProtectKostra}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{CalibrateSSB}{\code{\link[CalibrateSSB]{AkuData}}}

  \item{easySdcTable}{\code{\link[easySdcTable]{EasyData}}}

  \item{Kostra}{\code{\link[Kostra]{ProtectKostra}}}

  \item{SmallCountRounding}{\code{\link[SmallCountRounding]{SmallCountData}}}

  \item{SSBtools}{\code{\link[SSBtools]{HierarchyCompute}}, \code{\link[SSBtools]{SSBtoolsData}}}
}}

